Chore/5-data-collect: 패키지 구조 변경
- 패키지 이름 변경
    - data -> collect
    - collect -> openApi
- 공통 패키지 이동
    - /domain/common -> /global/common
