+++ docker-compose.yml
@@ -38,7 +38,7 @@ services:
     ports:
       - "27017:27017"
     volumes:
-      - mongo_data:/data/db
+      - mongo_data:/collect/db
       - ./mongo-init:/docker-entrypoint-initdb.d
     networks:
       - my-network
