+++ src/main/java/com/likelion/cleopatra/domain/collect/exception/LinkCollectException.java
@@ -1,0 +1,7 @@
+package com.likelion.cleopatra.domain.collect.exception;
+
+import com.likelion.cleopatra.global.exception.CleopatraException;
+
+public class LinkCollectException extends CleopatraException {
+    public LinkCollectException(LinkCollectErrorCode errorCode) { super(errorCode); }
+}
