--- src/main/java/com/likelion/cleopatra/domain/data/exception/LinkCollectException.java
@@ -1,7 +1,0 @@
-package com.likelion.cleopatra.domain.data.exception;
-
-import com.likelion.cleopatra.global.exception.CleopatraException;
-
-public class LinkCollectException extends CleopatraException {
-    public LinkCollectException(LinkCollectErrorCode errorCode) { super(errorCode); }
-}
