+++ src/main/java/com/likelion/cleopatra/domain/openApi/naver/dto/blog/NaverBlogSearchRes.java
@@ -1,0 +1,16 @@
+package com.likelion.cleopatra.domain.openApi.naver.dto.blog;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+public class NaverBlogSearchRes {
+    private String lastBuildDate;
+    private int total;
+    private int start;
+    private int display;
+    private List<NaverBlogItem> items;
+}
