--- src/main/java/com/likelion/cleopatra/domain/platform/exception/OpenApiException.java
@@ -1,7 +1,0 @@
-package com.likelion.cleopatra.domain.platform.exception;
-
-import com.likelion.cleopatra.global.exception.CleopatraException;
-
-public class OpenApiException extends CleopatraException {
-    public OpenApiException(OpenApiErrorCode errorCode) { super(errorCode); }
-}
